make default export for Box
