+++ src/components/Box/Box.mdx
@@ -6,7 +6,7 @@ route: /components/box
 
 import { PropsTable } from "docz";
 import { ExampleContainer } from "../../docs/helpers.ts";
-import { Box } from "./Box.tsx";
+import Box from ".";
 
 # Box
